Update and rename exe_3.c to exe_3.cs

+++ exe_3.cs
@@ -14,7 +14,7 @@ namespace Exercicios_aula
             int  altura = Convert.ToInt32(Console.ReadLine());
     
 
-            int calcularArea =  base1 * altura;
+            double calcularArea =  base1 * altura;
